:#53 Moving Event factories to separate namespace
With adding all other GitHub events, Event namespace will get crowded
quickly.
