+++ spec/EventFactory/PushEventFactorySpec.php
@@ -2,17 +2,17 @@
 
 declare(strict_types=1);
 
-namespace spec\Devboard\GitHub\Webhook\Core\Event;
+namespace spec\Devboard\GitHub\Webhook\Core\EventFactory;
 
 use Devboard\GitHub\GitHubCommit;
 use Devboard\GitHub\GitHubRepo;
 use Devboard\GitHub\Webhook\Core\Commit\GitHubCommitFactory;
 use Devboard\GitHub\Webhook\Core\Event\PushEvent;
 use Devboard\GitHub\Webhook\Core\Event\PushEvent\Pusher;
-use Devboard\GitHub\Webhook\Core\Event\PushEvent\PusherFactory;
-use Devboard\GitHub\Webhook\Core\Event\PushEventFactory;
 use Devboard\GitHub\Webhook\Core\Event\Sender;
-use Devboard\GitHub\Webhook\Core\Event\SenderFactory;
+use Devboard\GitHub\Webhook\Core\EventFactory\PushEvent\PusherFactory;
+use Devboard\GitHub\Webhook\Core\EventFactory\PushEventFactory;
+use Devboard\GitHub\Webhook\Core\EventFactory\SenderFactory;
 use Devboard\GitHub\Webhook\Core\Repo\GitHubRepoFactory;
 use PhpSpec\ObjectBehavior;
