+++ src/EventFactory/PushEvent/PusherFactory.php
@@ -2,7 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Devboard\GitHub\Webhook\Core\Event\PushEvent;
+namespace Devboard\GitHub\Webhook\Core\EventFactory\PushEvent;
+
+use Devboard\GitHub\Webhook\Core\Event\PushEvent\Pusher;
 
 /**
  * @see PusherFactorySpec
